--- src/server/contact-poe/configuration/demandeDeContactPOEDependency.ts
@@ -1,14 +1,0 @@
-import { DemandeDeContactPOERepository } from '~/server/contact-poe/domain/DemandeDeContactPOERepository';
-import { EnvoyerDemandeDeContactPOEUsecase } from '~/server/contact-poe/usecase/envoyerDemandeDeContactPOE.usecase';
-
-export interface DemandeDeContactPOEDependency {
-  demandeDeContactPOEUsecase: EnvoyerDemandeDeContactPOEUsecase
-}
-
-export const contactpoeDependenciesContainer = (
-  demandeDeContactPOERepository: DemandeDeContactPOERepository,
-): DemandeDeContactPOEDependency => {
-  return {
-    demandeDeContactPOEUsecase: new EnvoyerDemandeDeContactPOEUsecase(demandeDeContactPOERepository),
-  };
-};
