+++ src/server/demande-de-contact/infra/strapiDemandeDeContact.repository.ts
@@ -2,7 +2,7 @@ import { createFailure, createSuccess, Either } from '~/server/errors/either';
 import { ErreurMétier } from '~/server/errors/erreurMétier.types';
 import { HttpClientServiceWithAuthentification } from '~/server/services/http/httpClientWithAuthentification.service';
 
-import { DemandeDeContactCEJ, DemandeDeContactEntreprise } from '../domain/DemandeDeContact';
+import { DemandeDeContactCEJ, DemandeDeContactEntreprise, DemandeDeContactPOE } from '../domain/DemandeDeContact';
 import { DemandeDeContactRepository } from '../domain/DemandeDeContact.repository';
 
 export class StrapiDemandeDeContactRepository implements DemandeDeContactRepository {
@@ -39,6 +39,32 @@ export class StrapiDemandeDeContactRepository implements DemandeDeContactReposit
           prenom: demandeDeContactEntreprise.prénom,
           sujet: demandeDeContactEntreprise.sujet,
           telephone: demandeDeContactEntreprise.téléphone,
+        },
+      });
+    } catch (error) {
+      return createFailure(ErreurMétier.SERVICE_INDISPONIBLE);
+    }
+    return createSuccess(undefined);
+  }
+
+  async savePOE(demandeDeContactPOE: DemandeDeContactPOE, annotation?: string): Promise<Either<void>> {
+    try {
+      await this.httpClientServiceWithAuthentification.post('contacts-poe', {
+        data: {
+          code_postal: demandeDeContactPOE.codePostal,
+          commentaire: demandeDeContactPOE.commentaire,
+          email: demandeDeContactPOE.email,
+          nom: demandeDeContactPOE.nom,
+          nom_societe: demandeDeContactPOE.nomSociété,
+          nombre_a_recruter: demandeDeContactPOE.nombreARecruter,
+          prenom: demandeDeContactPOE.prénom,
+          secteur: demandeDeContactPOE.secteur,
+          siret: demandeDeContactPOE.siret,
+          taille: demandeDeContactPOE.taille,
+          telephone: demandeDeContactPOE.téléphone,
+          travail: demandeDeContactPOE.travail,
+          ville: demandeDeContactPOE.ville,
+          ...(annotation ? { erreur: annotation } : {}),
         },
       });
     } catch (error) {
